refactor!: move mls bundles to a unified module tree
We no longer implement manual getters for each member here in Wasm;
we instead use the `getter_with_clone` adjustment. This is a small
incompatibility, but it's an incompatibility. More in this vein will come.
