+++ crypto-ffi/src/lib.rs
@@ -33,6 +33,7 @@ pub mod bindings;
 #[cfg(not(target_family = "wasm"))]
 uniffi::setup_scaffolding!("core_crypto_ffi");
 
+mod bundles;
 mod ciphersuite;
 mod client_id;
 mod credential_type;
@@ -40,6 +41,10 @@ mod crl;
 mod error;
 mod metadata;
 
+pub use bundles::{
+    commit::CommitBundle, group_info::GroupInfoBundle, proposal::ProposalBundle,
+    proteus_auto_prekey::ProteusAutoPrekeyBundle, welcome::WelcomeBundle,
+};
 #[cfg(target_family = "wasm")]
 pub(crate) use ciphersuite::lower_ciphersuites;
 pub use ciphersuite::{Ciphersuite, Ciphersuites};
