+++ crypto-ffi/src/wasm/mod.rs
@@ -26,174 +26,9 @@ use wasm_bindgen::{JsCast, prelude::*};
 use wasm_bindgen_futures::future_to_promise;
 
 use crate::{
-    Ciphersuite, CoreCryptoError, CredentialType, FfiClientId, InternalError, MlsError, WasmCryptoResult,
-    lower_ciphersuites,
+    Ciphersuite, CommitBundle, CoreCryptoError, CredentialType, FfiClientId, InternalError, MlsError, ProposalBundle,
+    WasmCryptoResult, lower_ciphersuites,
 };
-
-#[wasm_bindgen]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct ProteusAutoPrekeyBundle {
-    pub id: u16,
-    #[wasm_bindgen(getter_with_clone)]
-    pub pkb: Vec<u8>,
-}
-
-#[wasm_bindgen]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct CommitBundle {
-    commit: Vec<u8>,
-    welcome: Option<Vec<u8>>,
-    group_info: GroupInfoBundle,
-}
-
-#[wasm_bindgen]
-impl CommitBundle {
-    #[wasm_bindgen(getter)]
-    pub fn commit(&self) -> Uint8Array {
-        Uint8Array::from(&*self.commit)
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn welcome(&self) -> Option<Uint8Array> {
-        self.welcome.as_ref().map(|buf| Uint8Array::from(buf.as_slice()))
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn group_info(&self) -> GroupInfoBundle {
-        self.group_info.clone()
-    }
-}
-
-impl TryFrom<MlsCommitBundle> for CommitBundle {
-    type Error = CoreCryptoError;
-
-    fn try_from(msg: MlsCommitBundle) -> Result<Self, Self::Error> {
-        let (welcome, commit, pgs) = msg.to_bytes_triple()?;
-
-        Ok(Self {
-            welcome,
-            commit,
-            group_info: pgs.into(),
-        })
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct GroupInfoBundle {
-    encryption_type: u8,
-    ratchet_tree_type: u8,
-    payload: Vec<u8>,
-}
-
-#[wasm_bindgen]
-impl GroupInfoBundle {
-    #[wasm_bindgen(getter)]
-    pub fn encryption_type(&self) -> u8 {
-        self.encryption_type
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn ratchet_tree_type(&self) -> u8 {
-        self.ratchet_tree_type
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn payload(&self) -> Uint8Array {
-        Uint8Array::from(&*self.payload)
-    }
-}
-
-impl From<MlsGroupInfoBundle> for GroupInfoBundle {
-    fn from(gi: MlsGroupInfoBundle) -> Self {
-        Self {
-            encryption_type: gi.encryption_type as u8,
-            ratchet_tree_type: gi.ratchet_tree_type as u8,
-            payload: gi.payload.bytes(),
-        }
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct ProposalBundle {
-    /// TLS-serialized MLS proposal that needs to be fanned out to other (existing) members of the conversation
-    proposal: Vec<u8>,
-    /// Unique identifier of a proposal.
-    proposal_ref: Vec<u8>,
-    /// New CRL Distribution of members of this group
-    crl_new_distribution_points: Option<Vec<String>>,
-}
-
-#[wasm_bindgen]
-impl ProposalBundle {
-    #[wasm_bindgen(getter)]
-    pub fn proposal(&self) -> Uint8Array {
-        Uint8Array::from(&*self.proposal)
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn proposal_ref(&self) -> Uint8Array {
-        Uint8Array::from(&*self.proposal_ref)
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn crl_new_distribution_points(&self) -> Option<js_sys::Array> {
-        self.crl_new_distribution_points
-            .clone()
-            .map(|crl_dp| crl_dp.iter().cloned().map(JsValue::from).collect::<js_sys::Array>())
-    }
-}
-
-impl TryFrom<MlsProposalBundle> for ProposalBundle {
-    type Error = CoreCryptoError;
-
-    fn try_from(msg: MlsProposalBundle) -> Result<Self, Self::Error> {
-        let (proposal, proposal_ref, crl_new_distribution_points) = msg.to_bytes()?;
-
-        Ok(Self {
-            proposal,
-            proposal_ref,
-            crl_new_distribution_points: crl_new_distribution_points.into(),
-        })
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct WelcomeBundle {
-    /// Identifier of the joined conversation
-    id: ConversationId,
-    /// New CRL Distribution of members of this group
-    crl_new_distribution_points: Option<Vec<String>>,
-}
-
-#[wasm_bindgen]
-impl WelcomeBundle {
-    /// Identifier of the joined conversation
-    #[wasm_bindgen(getter)]
-    pub fn id(&self) -> Uint8Array {
-        Uint8Array::from(&*self.id)
-    }
-
-    /// New CRL Distribution of members of this group
-    #[wasm_bindgen(getter)]
-    #[wasm_bindgen(js_name = "crlNewDistributionPoints")]
-    pub fn crl_new_distribution_points(&self) -> Option<js_sys::Array> {
-        self.crl_new_distribution_points
-            .clone()
-            .map(|crl_dp| crl_dp.iter().cloned().map(JsValue::from).collect::<js_sys::Array>())
-    }
-}
-
-impl From<core_crypto::prelude::WelcomeBundle> for WelcomeBundle {
-    fn from(w: core_crypto::prelude::WelcomeBundle) -> Self {
-        Self {
-            id: w.id,
-            crl_new_distribution_points: w.crl_new_distribution_points.into(),
-        }
-    }
-}
 
 #[wasm_bindgen]
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
